% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COL.R
\name{conserve}
\alias{conserve}
\title{conserve}
\usage{
conserve(
  cirRNA,
  STAR_output,
  genome1 = "Hgenome",
  genomeOther1,
  genome2 = "Mgenome",
  genomeOther2,
  chainObject,
  outpath
)
}
\arguments{
\item{cirRNA}{cirularRNA data}

\item{STAR_output}{output file from STAR}

\item{genome1}{DNA sequence (FASTA), by default Homo sapiens}

\item{genomeOther1}{other DNA sequence (FASTA) instead of Homo sapiens}

\item{genome2}{DNA sequence (FASTA), by default Mus musculus}

\item{genomeOther2}{other DNA sequence (FASTA) instead of Mus musculus}

\item{chainObject}{chainObject usually imported using \code{\link[rtracklayer]{import.chain}} and downloaded data from https://hgdownload.soe.ucsc.edu/goldenPath/hg38/liftOver/}

\item{outpath}{output path}
}
\value{
Conserved motif result
}
\description{
conserve function is built to identify putatively functional back-splicing based on the molecular error hypothesis.COL combined (1) substantially higher back-splicing rates than expected from the total splicing amounts; (2) conserved splicing motifs between different species; (3) high back-splicing levels to identify functional back-splicing.
}
\seealso{
position specific liftover function: \code{\link[COL]{liftOVER}} , outlier identification function: \code{\link[COL]{outlier}} , Moftif identification function: \code{\link[COL]{cmotif}}
}
