% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COL.R
\name{genome}
\alias{genome}
\title{genome}
\usage{
genome(genome.UCSC, genomeOther)
}
\arguments{
\item{genome.UCSC}{genome.UCSC DNA sequence (FASTA format); for example Hgenome for Homo sapiens and Mgenome for Mus musculus}

\item{genomeOther}{other DNA sequence (FASTA) instead of Homo sapiens; if user use independent genome FASTA where parameter #genome.UCSC="others".}
}
\description{
Genome sequences for Homo sapiens (see more \code{\link[BSgenome.Hsapiens.UCSC.hg38]{BSgenome.Hsapiens.UCSC.hg38}}) and Mus musculus (see more \code{\link[BSgenome.Mmusculus.UCSC.mm10]{BSgenome.Mmusculus.UCSC.mm10}})
}
