% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COL.R
\name{liftOVER}
\alias{liftOVER}
\title{liftOVER}
\usage{
liftOVER(output, chainObject)
}
\arguments{
\item{output}{circular RNA data, usually data have chromosome,start,end and so on,
liftOVER function accept the data format like chr,start,end,strand and so on}

\item{chainObject}{chainObject usually imported using \code{\link[rtracklayer]{import.chain}} and downloaded data from https://hgdownload.soe.ucsc.edu/goldenPath/hg38/liftOver/}
}
\value{
lifting result with chromosome, start, end, strand and so on
}
\description{
liftOVER function is built based on the UCSC liftover tools, which are designed for lifting features from specific region rather than
an interval region from one genome to other.
}
\references{
https://genome.ucsc.edu/cgi-bin/hgLiftOver
}
\seealso{
\code{\link[rtracklayer]{liftOver}} for more details
}
