% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COL.R
\name{genome}
\alias{genome}
\title{genome}
\usage{
genome(genome.UCSC, genomeOther)
}
\arguments{
\item{genome.UCSC}{DNA sequence (FASTA) Hgenome:BSgenome.Hsapiens.UCSC.hg38; Mgenome: BSgenome.Mmusculus.UCSC.mm10}

\item{genomeOther}{other DNA sequence (FASTA) instead of Homo sapiens; if user use independent genome FASTA.}
}
\description{
COL:a pipeline for identifying putatively functional back-splicing
}
\details{
COL aims to identify putatively functional back-splicing based on the molecular error hypothesis.
COL combined (1) substantially higher back-splicing rates than expected from the total splicing amounts;
(2) conserved splicing motifs between different species;
(3) high back-splicing levels to identify functional back-splicing.
COL has been proved to be very effective on identifying shared back-splicing junctions and experimentally functional back-splicing junctions.

Dependencies pacakage
}
