% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COL.R
\name{outlier}
\alias{outlier}
\title{outlier}
\usage{
outlier(cirRNA, STAR_output, k = 4)
}
\arguments{
\item{cirRNA}{cirularRNA data}

\item{STAR_output}{output file from STAR}

\item{k}{the Cook's distance cutoff for outliers}
}
\description{
outlier function used to detect outliers based on COOKs distance
}
\details{
Dependencies pacakage
}
